test: Cars 클래스에 대한 테스트 클래스를 별도로 생성한다.
